[SCMOD-7476] Add a log4j2 caf logging component
