+++ .gitignore
@@ -1,1 +1,5 @@
 target/
+
+.settings
+.project
+.classpath
