+++ caf-logging-common/src/main/java/com/github/cafapi/logging/common/JsonFactorCreator.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2019 Micro Focus or one of its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.cafapi.logging.common;
+
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonGenerator.Feature;
+
+
+public final class JsonFactorCreator
+{    
+    public static JsonFactory createJsonFactory()
+    {
+        final JsonFactory newJsonFactory = new JsonFactory();
+        newJsonFactory.disable(Feature.AUTO_CLOSE_JSON_CONTENT);
+        newJsonFactory.disable(Feature.AUTO_CLOSE_TARGET);
+        newJsonFactory.disable(Feature.FLUSH_PASSED_TO_STREAM);
+
+        return newJsonFactory;
+    }
+}
