+++ caf-logging-common/src/main/java/com/github/cafapi/logging/common/UnexpectedIOException.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2019 Micro Focus or one of its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.cafapi.logging.common;
+
+import java.io.IOException;
+
+public final class UnexpectedIOException extends RuntimeException
+{
+    private static final long serialVersionUID = -3409099815468419997L;
+
+    public UnexpectedIOException(final IOException ex)
+    {
+        super(ex);
+    }
+}
