+++ caf-logging-log4j2/src/main/resources/log4j2.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright 2019 Micro Focus or one of its affiliates.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+         http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+
+<Configuration>
+    <Appenders>
+        <Console name="STDOUT" target="SYSTEM_OUT">
+            <PatternLayout
+            pattern="[%d{HH:mm:ss.SSS}{UTC}Z %aProcessId %-5level ${%-12.-12replace{ %X{tenantId} }{[^\w\-\.]}{}:\-} ${%-4.-4sanitize{ %X{correlationId} }:\-}] %sanitize{ %logger{30} }: %maybeJsonMsgAndEx%n" 
+            charset="UTF-8"/>
+        </Console>
+    </Appenders>
+    <Loggers>
+        <Root level="${env:CAF_LOG_LEVEL:-INFO}">
+            <AppenderRef ref="STDOUT"/>
+        </Root> 
+    </Loggers>
+</Configuration>
