+++ deptree.log
@@ -1,0 +1,61 @@
+[INFO] Scanning for projects...
+[INFO] Inspecting build with total of 3 modules...
+[INFO] Installing Nexus Staging features:
+[INFO]   ... total of 3 executions of maven-deploy-plugin replaced with nexus-staging-maven-plugin
+[INFO] ------------------------------------------------------------------------
+[INFO] Reactor Build Order:
+[INFO] 
+[INFO] caf-logging
+[INFO] caf-logging-log4j2
+[INFO] caf-logging-logback
+[INFO] 
+[INFO] ------------------------------------------------------------------------
+[INFO] Building caf-logging 1.0.0-SNAPSHOT
+[INFO] ------------------------------------------------------------------------
+[INFO] 
+[INFO] --- maven-dependency-plugin:3.1.1:tree (default-cli) @ caf-logging ---
+[INFO] com.github.cafapi:caf-logging:pom:1.0.0-SNAPSHOT
+[INFO] 
+[INFO] ------------------------------------------------------------------------
+[INFO] Building caf-logging-log4j2 1.0.0-SNAPSHOT
+[INFO] ------------------------------------------------------------------------
+[INFO] 
+[INFO] --- maven-dependency-plugin:3.1.1:tree (default-cli) @ caf-logging-log4j2 ---
+[INFO] com.github.cafapi:caf-logging-log4j2:jar:1.0.0-SNAPSHOT
+[INFO] +- org.apache.logging.log4j:log4j-api:jar:2.11.2:compile
+[INFO] +- org.apache.logging.log4j:log4j-core:jar:2.11.2:compile
+[INFO] +- com.github.cafapi.util:util-process-identifier:jar:1.16.0-118:compile
+[INFO] +- com.fasterxml.jackson.core:jackson-core:jar:2.9.8:compile
+[INFO] +- org.apache.commons:commons-text:jar:1.2:compile
+[INFO] |  \- org.apache.commons:commons-lang3:jar:3.7:compile
+[INFO] +- junit:junit:jar:4.12:test
+[INFO] \- org.hamcrest:hamcrest-core:jar:1.3:test
+[INFO] 
+[INFO] ------------------------------------------------------------------------
+[INFO] Building caf-logging-logback 1.0.0-SNAPSHOT
+[INFO] ------------------------------------------------------------------------
+[INFO] 
+[INFO] --- maven-dependency-plugin:3.1.1:tree (default-cli) @ caf-logging-logback ---
+[INFO] com.github.cafapi:caf-logging-logback:jar:1.0.0-SNAPSHOT
+[INFO] +- ch.qos.logback:logback-classic:jar:1.2.3:compile
+[INFO] +- ch.qos.logback:logback-core:jar:1.2.3:compile
+[INFO] +- com.github.cafapi.util:util-process-identifier:jar:1.16.0-118:compile
+[INFO] +- com.fasterxml.jackson.core:jackson-core:jar:2.9.8:compile
+[INFO] +- org.apache.commons:commons-text:jar:1.2:compile
+[INFO] |  \- org.apache.commons:commons-lang3:jar:3.7:compile
+[INFO] +- junit:junit:jar:4.12:test
+[INFO] +- org.hamcrest:hamcrest-core:jar:1.3:test
+[INFO] \- org.slf4j:slf4j-api:jar:1.7.25:test
+[INFO] ------------------------------------------------------------------------
+[INFO] Reactor Summary:
+[INFO] 
+[INFO] caf-logging ........................................ SUCCESS [  2.273 s]
+[INFO] caf-logging-log4j2 ................................. SUCCESS [  0.130 s]
+[INFO] caf-logging-logback ................................ SUCCESS [  0.052 s]
+[INFO] ------------------------------------------------------------------------
+[INFO] BUILD SUCCESS
+[INFO] ------------------------------------------------------------------------
+[INFO] Total time: 4.648 s
+[INFO] Finished at: 2019-10-17T23:05:50-07:00
+[INFO] Final Memory: 19M/491M
+[INFO] ------------------------------------------------------------------------
