--- src/main/resources/logback.xml
@@ -1,37 +1,0 @@
-<!--
-
-    Copyright 2019 Micro Focus or one of its affiliates.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-         http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<configuration>
-    <conversionRule conversionWord="sanitize"
-                    converterClass="com.github.cafapi.logging.SanitizeConverter" />
-    <conversionRule conversionWord="maybeJsonMsgAndEx"
-                    converterClass="com.github.cafapi.logging.MaybeJsonMessageConverter" />
-
-    <appender name="STDOUT" class="com.github.cafapi.logging.CAFConsoleAppender">
-        <encoder> 
-            <pattern>
-                [%date{HH:mm:ss.SSS,UTC}Z %thread %-5level %-12.-12sanitize(%mdc{tenantId:--}) %-4.-4sanitize(%mdc{correlationId:--})] %sanitize(%logger{30}): %maybeJsonMsgAndEx%n
-            </pattern>
-            <charset>UTF-8</charset>
-        </encoder>
-    </appender>
-
-    <root level="${CAF_LOG_LEVEL:-INFO}">
-        <appender-ref ref="STDOUT" />
-    </root>
-    <logger name="org.hibernate.engine" level="${CAF_LOG_LEVEL_HIBERNATE:-ERROR}" />
-</configuration>
